Seperated Calendar into different modules. Fixes #33
